requirements.txt
aiofiles>=24.1.0
aiohttp>=3.11.0
aiosqlite>=0.21.0
magic-filter>=1.0.12
maxapi
openpyxl>=3.1.0
pydantic>=2.11.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
loguru>=0.7.0
